Give children their own piece of state.
Seemed a bit confusing if some properties of an object have metadata
about fetching status or errors. Better keep separate to reflect the
api structure.
