+++ src/reducers/index.js
@@ -6,6 +6,7 @@ import metadata from './metadata'
 import session from './session'
 import entityCache from './entityCache';
 import documentCache from './documentCache';
+import documentChildrenResults from './documentChildrenResults';
 
 const rootReducer = combineReducers({
   collections,
@@ -14,6 +15,7 @@ const rootReducer = combineReducers({
   session,
   entityCache,
   documentCache,
+  documentChildrenResults,
 })
 
 export default rootReducer;
